# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cape Christian Sermon Planning API
  description: RESTful API for sermon planning and management system
  version: 1.0.0
  contact:
    name: Cape Christian Tech Team
    email: tech@capechristian.org
servers:
  - url: https://api.sermons.capechristian.org/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /series:
    get:
      tags: [Series]
      summary: List sermon series
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, active, completed, archived]
      responses:
        200:
          description: List of sermon series
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SermonSeries'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Series]
      summary: Create new sermon series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SermonSeriesCreate'
      responses:
        201:
          description: Series created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SermonSeries'
  /series/{id}:
    get:
      tags: [Series]
      summary: Get series by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        200:
          description: Series details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SermonSeries'

    put:
      tags: [Series]
      summary: Update series
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SermonSeriesUpdate'
      responses:
        200:
          description: Series updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SermonSeries'

    delete:
      tags: [Series]
      summary: Delete series
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        204:
          description: Series deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    SermonSeries:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        theme:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        seriesNumber:
          type: integer
        artworkUrl:
          type: string
        keyScriptures:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        targetAudience:
          type: string
        goals:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [planning, active, completed, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SermonSeriesCreate:
      type: object
      required: [title, startDate, endDate]
      properties:
        title:
          type: string
        subtitle:
          type: string
        theme:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        seriesNumber:
          type: integer
        keyScriptures:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        targetAudience:
          type: string
        goals:
          type: array
          items:
            type: string
    SermonSeriesUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        theme:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        artworkUrl:
          type: string
        keyScriptures:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        targetAudience:
          type: string
        goals:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [planning, active, completed, archived]

    # Sermon Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...